+++ packages/snippets/src/browser/addElementIds.js
@@ -1,4 +1,15 @@
 function addElementIds([elements, ids]) {
+  if (!elements) return []
+  let elementsFromSelectors = []
+
+  elements.map(element => {
+    if (typeof element === 'string') {
+      const curr = document.querySelectorAll(element)
+      elementsFromSelectors = elementsFromSelectors.concat(Object.values(curr))
+    }
+  })
+  elements = elementsFromSelectors.length > 0 ? elementsFromSelectors : elements
+
   return elements.map((element, index) => {
     const path = [element]
     if (element.getRootNode) {
@@ -13,7 +24,7 @@ function addElementIds([elements, ids]) {
       const oldElementId = element.getAttribute('data-applitools-selector')
       const newElementId = oldElementId ? `${oldElementId} ${elementId}` : elementId
       element.setAttribute('data-applitools-selector', newElementId)
-      return `[data-applitools-selector~="${elementId}"]`
+      return `[data-applitools-selector~='${elementId}']`
     })
   })
 }